+++ @liminal/sqlite/.npmignore
