+++ @liminal/sqlite/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "liminal-sqlite",
+  "name": "@liminal/sqlite",
   "version": "0.0.0",
   "repository": {
     "type": "git",
@@ -17,7 +17,7 @@
       "bun": "./schemas/index.ts"
     }
   },
-  "description": "Postgres adapter for Liminal.",
+  "description": "SQLite storage adapter for Liminal threads.",
   "license": "Apache-2.0",
   "publishConfig": {
     "access": "public",
@@ -26,7 +26,6 @@
   "sideEffects": false,
   "type": "module",
   "peerDependencies": {
-    "@effect/ai": "^0.25.2",
     "@effect/sql-drizzle": "^0.43.0",
     "drizzle-orm": "^0.44.4",
     "effect": "^3.17.7",
