+++ @liminal/sqlite/tables/index.ts
