+++ examples/package.json
@@ -15,6 +15,6 @@
     "@electric-sql/pglite": "^0.3.7",
     "effect": "^3.17.5",
     "liminal": "workspace:*",
-    "liminal-sqlite": "workspace:*"
+    "@liminal/sqlite": "workspace:*"
   }
 }
