+++ examples/tsconfig.json
@@ -12,6 +12,6 @@
   "exclude": [],
   "references": [
     { "path": "../liminal" },
-    { "path": "../liminal-sqlite" }
+    { "path": "../@liminal/sqlite" }
   ]
 }
