+++ lmnl.im/client/package.json
@@ -1,12 +1,12 @@
 {
-  "name": "@liminal.im/client",
+  "name": "@lmnl.im/client",
   "private": true,
   "type": "module",
   "dependencies": {
     "@clerk/clerk-react": "^5.40.0",
     "@effect-atom/atom-react": "^0.1.5",
     "@effect/platform": "^0.90.1",
-    "@liminal.im/domain": "workspace:*",
+    "@lmnl.im/domain": "workspace:*",
     "clsx": "^2.1.1",
     "effect": "^3.17.7",
     "react": "^19.1.1",
