+++ lmnl.im/domain/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@liminal.im/domain",
+  "name": "@lmnl.im/domain",
   "private": true,
   "type": "module",
   "dependencies": {
